@@ -6,13 +6,12 @@
     "homepage": "https://github.com/wp-media/wp-rocket-cli",
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.6"
     },
     "authors": [{
         "name": "WP Rocket Team",
-        "homepage": "http://wp-rocket.me",
-        "email": "jonathan@wp-rocket.me",
-        "role": "Developer"
+        "homepage": "https://wp-rocket.me",
+        "email": "contact@wp-rocket.me"
     }],
     "support": {
 		"issues": "https://github.com/wp-media/wp-rocket-cli/issues",
