@@ -1,4 +1,8 @@
 <?php
+
+use function WP_CLI\Utils\make_progress_bar;
+use WP_Rocket\Engine\Cache\WPCache;
+
 /**
  * Manage Revisions
  */
@@ -17,11 +21,12 @@ class WPRocket_CLI extends WP_CLI_Command {
 
 		if( defined( 'WP_CACHE' ) && ! WP_CACHE ) {
 
-			if( is_writable( rocket_find_wpconfig_path() ) ) {
-				set_rocket_wp_cache_define( true );
+			$wp_cache = new WPCache( rocket_direct_filesystem() );
+
+			if ( $wp_cache->set_wp_cache_constant( true ) ) {
 				WP_CLI::success( 'WP Rocket is now enabled, WP_CACHE is set to true.' );
 			} else {
-				WP_CLI::error( 'It seems we don\'t have writing permissions on wp-config.php file.' );
+				WP_CLI::error( 'Error while setting WP_CACHE constant into wp-config.php!' );
 			}
 
 		} else {
@@ -43,11 +48,12 @@ class WPRocket_CLI extends WP_CLI_Command {
 
 		if( defined( 'WP_CACHE' ) && WP_CACHE ) {
 
-			if( is_writable( rocket_find_wpconfig_path() ) ) {
-				set_rocket_wp_cache_define( false );
+			$wp_cache = new WPCache( rocket_direct_filesystem() );
+
+			if ( $wp_cache->set_wp_cache_constant( false ) ) {
 				WP_CLI::success( 'WP Rocket is now disabled, WP_CACHE is set to false.' );
 			} else {
-				WP_CLI::error( 'It seems we don\'t have writing permissions on wp-config.php file.' );
+				WP_CLI::error( 'Error while setting WP_CACHE constant into wp-config.php!' );
 			}
 
 		} else {
@@ -90,24 +96,23 @@ class WPRocket_CLI extends WP_CLI_Command {
 	 *
 	 * @subcommand clean
 	 */
-	public function clean( $args = array(), $assoc_args = array() ) {
+	public function clean( array $args = [], array $assoc_args = [] ) {
 		if ( ! function_exists( 'rocket_clean_domain' ) ) {
 			WP_CLI::error( ' The plugin WP-Rocket seems not enabled on this site.' );
 		}
 
 		if( ! empty( $assoc_args['blog_id'] ) ) {
-
 			if ( ! defined( 'MULTISITE' ) || ! MULTISITE ) {
 				WP_CLI::error( 'This installation doesn\'t support multisite.' );
 			}
 
 			$blog_ids = explode( ',' , $assoc_args['blog_id'] );
-			$blog_ids = array_map( 'trim' , $blog_ids );
 			$total    = 0;
 
 			$notify = \WP_CLI\Utils\make_progress_bar( 'Delete cache files', count( $blog_ids ) );
 
 			foreach ( $blog_ids as $blog_id ) {
+				$blog_id = trim( $blog_id );
 
 				if ( $bloginfo = get_blog_details( (int) $blog_id, false ) ) {
 
@@ -119,7 +124,6 @@ class WPRocket_CLI extends WP_CLI_Command {
 					restore_current_blog();
 
 					$total++;
-
 				} else {
 					WP_CLI::line( 'This blog ID "' . $blog_id . '" doesn\'t exist.' );
 				}
@@ -132,9 +136,8 @@ class WPRocket_CLI extends WP_CLI_Command {
 			WP_CLI::success( 'Cache cleared for ' . $total . ' blog(s).' );
 
 		} else if( ! empty( $assoc_args['lang'] ) ) {
-
 			if( ! rocket_has_i18n() ) {
-				WP_CLI::error( 'No WPML or qTranslate in this website.' );
+				WP_CLI::error( 'No WPML, Polylang or qTranslate in this website.' );
 			}
 
 			$langs = explode( ',' , $assoc_args['lang'] );
@@ -145,7 +148,7 @@ class WPRocket_CLI extends WP_CLI_Command {
 
 			foreach ( $langs as $lang ) {
 
-				rocket_clean_domain_for_selected_lang( $lang );
+				rocket_clean_domain( $lang );
 				$notify->tick();
 
 			}
@@ -154,56 +157,37 @@ class WPRocket_CLI extends WP_CLI_Command {
 			WP_CLI::success( 'Cache files cleared for ' . $total . ' lang(s).' );
 
 		} else if( ! empty( $assoc_args['permalink'] ) ) {
-
 			$permalinks = explode( ',' , $assoc_args['permalink'] );
-			$permalinks = array_map( 'trim' , $permalinks );
 			$total      = count( $permalinks );
 
-			$notify = \WP_CLI\Utils\make_progress_bar( 'Delete cache files', $total );
+			$notify = make_progress_bar( 'Delete cache files', $total );
 
 			foreach ( $permalinks as $permalink ) {
+				$permalink = trim( $permalink );
 
 				rocket_clean_files( $permalink );
 				WP_CLI::line( 'Cache cleared for "' . $permalink . '".' );
 
 				$notify->tick();
-
 			}
 
 			$notify->finish();
 			WP_CLI::success( 'Cache files cleared for ' . $total . ' permalink(s).' );
 
 		} else if( ! empty( $assoc_args['post_id'] ) ) {
-
 			$total    = 0;
 			$post_ids = explode( ',' , $assoc_args['post_id'] );
-			$post_ids = array_map( 'trim' , $post_ids );
 
-			$notify = \WP_CLI\Utils\make_progress_bar( 'Delete cache files', count( $post_ids ) );
+			$notify = make_progress_bar( 'Delete cache files', count( $post_ids ) );
 
 			foreach ( $post_ids as $post_id ) {
+				$post_id = trim( $post_id );
 
-				global $wpdb;
-				$post_exists = $wpdb->get_row( "SELECT ID FROM $wpdb->posts WHERE id = '" . (int) $post_id . "'");
-
-				if( $post_exists ) {
-
-					if( get_post_type( $post_id ) == 'attachment' ) {
-
-						WP_CLI::line( 'This post ID "' . $post_id . '" is an attachment.' );
-
-					} else {
-
-						rocket_clean_post( $post_id );
-						WP_CLI::line( 'Cache cleared for post ID "' . $post_id . '".' );
-						$total++;
-
-					}
-
+				if( rocket_clean_post( $post_id ) ) {
+					WP_CLI::line( 'Cache cleared for post ID "' . $post_id . '".' );
+					$total++;
 				} else {
-
-					WP_CLI::line( 'This post ID "' . $post_id . '" doesn\'t exist.' );
-
+					WP_CLI::line( 'This post ID "' . $post_id . '" is not a valid public post.' );
 				}
 
 				$notify->tick();
@@ -211,7 +195,6 @@ class WPRocket_CLI extends WP_CLI_Command {
 			}
 
 			if( $total ) {
-
 				$notify->finish();
 
 				if( $total == 1 ) {
@@ -221,20 +204,20 @@ class WPRocket_CLI extends WP_CLI_Command {
 				}
 
 			} else {
-				WP_CLI::error( 'No cache files are cleared.' );
+				WP_CLI::error( 'No cache files cleared.' );
 			}
-
 		} else {
-
 			if ( ! empty( $assoc_args['confirm'] ) && $assoc_args['confirm'] ) {
 				WP_CLI::line( 'Deleting all cache files.' );
 			} else {
 				WP_CLI::confirm( 'Delete all cache files ?' );
 			}
 
-			rocket_clean_domain();
-			WP_CLI::success( 'All cache files cleared.' );
-
+			if ( rocket_clean_domain() ) {
+				WP_CLI::success( 'All cache files cleared.' );
+			}else{
+				WP_CLI::error( 'No cache files are cleared.' );
+			}
 		}
 
 	}
@@ -246,7 +229,7 @@ class WPRocket_CLI extends WP_CLI_Command {
 	 *
 	 * [--sitemap]
 	 * : Trigger sitemap-based preloading
-	 * 
+	 *
 	 * ## EXAMPLES
 	 *
 	 *     wp rocket preload
@@ -254,14 +237,17 @@ class WPRocket_CLI extends WP_CLI_Command {
 	 *
 	 * @subcommand preload
 	 */
-	public function preload( $args = array(), $assoc_args = array() ) {
-	
+	public function preload( array $args = [], array $assoc_args = [] ) {
+
 		if ( ! empty( $assoc_args['sitemap'] ) && $assoc_args['sitemap'] ) {
 			WP_CLI::line( 'Triggering sitemap-based preloading.' );
 			run_rocket_sitemap_preload();
 		} else {
-			WP_CLI::line( 'Triggering homepage-based preloading.' );
-			run_rocket_bot( 'cache-preload' );
+			if ( run_rocket_bot() ) {
+				WP_CLI::success( 'Triggering homepage-based preloading.' );
+			}else{
+				WP_CLI::error( 'Cannot start preload cache, please check preload option is activated.' );
+			}
 		}
 
 		WP_CLI::success( 'Finished WP Rocket preload cache files.' );
@@ -288,32 +274,39 @@ class WPRocket_CLI extends WP_CLI_Command {
 	 *
 	 * @subcommand regenerate
 	 */
-	public function regenerate( $args = array(), $assoc_args = array() ) {
+	public function regenerate( array $args = [], array $assoc_args = [] ) {
 		if( !empty( $assoc_args['file'] ) ) {
 			switch( $assoc_args['file'] ) {
 				case 'advanced-cache':
-					rocket_generate_advanced_cache_file();
-					WP_CLI::success( 'The advanced-cache.php file has just been regenerated.' );
+					if ( rocket_generate_advanced_cache_file() ) {
+						WP_CLI::success( 'The advanced-cache.php file has just been regenerated.' );
+					}else{
+						WP_CLI::error( 'Cannot generate advanced-cache.php file, please check folder permissions.' );
+					}
 					break;
 				case 'config':
 					if ( ! empty( $assoc_args['nginx'] ) && $assoc_args = true ) {
 						$GLOBALS['is_nginx'] = true;
 					}
+
 					rocket_generate_config_file();
 					WP_CLI::success( 'The config file has just been regenerated.' );
 					break;
 				case 'htaccess':
 					$GLOBALS['is_apache'] = true;
-					flush_rocket_htaccess();
-					WP_CLI::success( 'The .htaccess file has just been regenerated.' );
+					if ( flush_rocket_htaccess() ) {
+						WP_CLI::success( 'The .htaccess file has just been regenerated.' );
+					}else{
+						WP_CLI::error( 'Cannot generate .htaccess file.' );
+					}
 					break;
 				default:
-					WP_CLI::error( 'You didn\'t specify a good value for the "file" argument. It should be: advanced-cache, config or htaccess.' );
+					WP_CLI::error( 'You did not specify a good value for the "file" argument. It should be: advanced-cache, config or htaccess.' );
 					break;
 			}
 
 		} else {
-			WP_CLI::error( 'You didn\'t specify the "file" argument.' );
+			WP_CLI::error( 'You did not specify the "file" argument.' );
 		}
 	}
 }
